Move ppc expd64/logd64 to common ieee754r
Keep a native decimal copy for all to use, and improve
as needed.
