+++ ieee754r/ddlogtbls.h
